requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Configuration management
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# API framework (for future steps)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # For testing FastAPI

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
ipython>=8.14.0
jupyter>=1.0.0

# Monitoring and logging
python-json-logger>=2.0.0

# Serialization
joblib>=1.3.0
boto3>=1.34.0
